pyproject.toml
[tool.poetry]
name = ""
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-swagger-ui = "^4.11.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
